Remove duplicate methods for 'submitForm' and 'editForm'
