+++ tests/e2e/default/enketo/add-family.wdio-spec.js
@@ -18,14 +18,14 @@ describe('Family form', () => {
     await commonPage.openFastActionReport('add-family-multiple-repeats', false);
     await familyForm.submitFamilyForm();
     await familyForm.reportCheck('test Family', 'boreholes', 'true', 'true', 'ucid');
-    await genericForm.editForm();
+    await reportsPage.editReport();
     await commonEnketoPage.setInputValue('Names', 'modified');
     await genericForm.nextPage(7);
     await commonEnketoPage.selectCheckBox('What is the family\'s source of drinking water?', 'Spring');
     await commonEnketoPage.selectRadioButton('Do they have mosquito nets', 'No');
     await commonEnketoPage.selectRadioButton('Do they have an hygienic toilet', 'No');
     await commonEnketoPage.selectCheckBox('Which method of Family Planning is being used', 'Condoms');
-    await reportsPage.submitForm();
+    await genericForm.submitForm();
     await familyForm.reportCheck('modified Family', 'boreholes spring', 'false', 'false', 'ucid condoms');
 
     const reportId = await reportsPage.getCurrentReportId();
