+++ tests/e2e/default/enketo/pregnancy-facility-visit-reminder.wdio-spec.js
@@ -48,7 +48,6 @@ describe('Health Facility ANC Reminder task', () => {
 
     await commonEnketoPage.selectRadioButton('Did you remind the client in-person or by phone?', 'In person');
     await genericForm.submitForm();
-    await commonPage.waitForPageLoaded();
 
     await commonPage.goToReports();
     await reportsPage.openSelectedReport(await reportsPage.firstReport());
