+++ tests/e2e/default/transitions/create-user-for-contacts.replace-user.wdio-spec.js
@@ -104,8 +104,7 @@ const populateReplaceUserForm = async (formID) => {
 
 const submitBasicForm = async () => {
   await commonPage.openFastActionReport('basic_form', false);
-  await reportsPage.submitForm();
-  await commonPage.waitForPageLoaded();
+  await genericForm.submitForm();
   return reportsPage.getCurrentReportId();
 };
 
@@ -216,7 +215,7 @@ describe('Create user for contacts', () => {
 
         await commonPage.goToPeople(originalContactId, true);
         await populateReplaceUserForm(REPLACE_USER_FORM_ID);
-        await contactsPage.submitForm();
+        await genericForm.submitForm();
         const reportNames = await contactsPage.getAllRHSReportsNames();
         expect(reportNames.filter(name => name === REPLACE_USER_FORM_TITLE)).to.have.lengthOf(1);
         await commonPage.goToReports();
@@ -346,7 +345,7 @@ describe('Create user for contacts', () => {
 
         await commonPage.goToPeople(originalContactId);
         await populateReplaceUserForm(REPLACE_USER_FORM_ID);
-        await contactsPage.submitForm(false);
+        await genericForm.submitForm();
 
         // Logout triggered immediately
         await (await loginPage.loginButton()).waitForDisplayed();
@@ -396,7 +395,7 @@ describe('Create user for contacts', () => {
 
         await commonPage.goToPeople(originalContactId);
         await populateReplaceUserForm(REPLACE_USER_FORM_ID);
-        await contactsPage.submitForm();
+        await genericForm.submitForm();
         let reportNames = await contactsPage.getAllRHSReportsNames();
         expect(reportNames.filter(name => name === REPLACE_USER_FORM_TITLE)).to.have.lengthOf(1);
         await commonPage.goToReports();
@@ -424,7 +423,7 @@ describe('Create user for contacts', () => {
         // Submit another replace user form
         await commonPage.goToPeople(replacementContactId0);
         await populateReplaceUserForm(OTHER_REPLACE_FORM_DOC.internalId);
-        await contactsPage.submitForm();
+        await genericForm.submitForm();
         reportNames = await contactsPage.getAllRHSReportsNames();
         expect(reportNames.filter(name => name === OTHER_REPLACE_FORM_DOC.title)).to.have.lengthOf(1);
         await commonPage.goToReports();
@@ -569,7 +568,7 @@ describe('Create user for contacts', () => {
 
         await commonPage.goToPeople(originalContactId);
         await populateReplaceUserForm(REPLACE_USER_FORM_ID);
-        await contactsPage.submitForm();
+        await genericForm.submitForm();
         const reportNames = await contactsPage.getAllRHSReportsNames();
         expect(reportNames.filter(name => name === REPLACE_USER_FORM_TITLE)).to.have.lengthOf(1);
         await commonPage.goToReports();
@@ -604,7 +603,7 @@ describe('Create user for contacts', () => {
         await commonPage.waitForPageLoaded();
         await commonPage.goToPeople(originalContactId);
         await populateReplaceUserForm(REPLACE_USER_FORM_ID);
-        await contactsPage.submitForm(false);
+        await genericForm.submitForm();
 
         // Logout triggered immediately
         await (await loginPage.loginButton()).waitForDisplayed();
@@ -679,7 +678,7 @@ describe('Create user for contacts', () => {
 
         await commonPage.goToPeople(originalContactId);
         await populateReplaceUserForm(REPLACE_USER_FORM_ID);
-        await contactsPage.submitForm();
+        await genericForm.submitForm();
         // No logout triggered
         await commonPage.goToReports();
         const reportId = await reportsPage.getLastSubmittedReportId();
@@ -715,7 +714,7 @@ describe('Create user for contacts', () => {
 
         await commonPage.goToPeople(originalContactId);
         await populateReplaceUserForm(REPLACE_USER_FORM_ID);
-        await contactsPage.submitForm();
+        await genericForm.submitForm();
         const reportNames = await contactsPage.getAllRHSReportsNames();
         expect(reportNames.filter(name => name === REPLACE_USER_FORM_TITLE)).to.have.lengthOf(1);
         await commonPage.goToReports();
@@ -792,7 +791,7 @@ describe('Create user for contacts', () => {
       await commonPage.goToPeople(originalContactId);
 
       await populateReplaceUserForm(REPLACE_USER_FORM_ID);
-      await contactsPage.submitForm();
+      await genericForm.submitForm();
       // No logout triggered
       await commonPage.goToReports();
       const reportId = await reportsPage.getLastSubmittedReportId();
